RL-I2IT/RL-I2IT-ImageStyleTransfer/requirements.txt
absl-py==0.13.0
altgraph==0.17
argon2-cffi==20.1.0
astunparse==1.6.3
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==3.3.1
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
clang==5.0
cloudpickle==1.6.0
colorama==0.4.4
cycler==0.10.0
debugpy==1.3.0
decorator==4.4.2
defusedxml==0.7.1
entrypoints==0.3
et-xmlfile==1.1.0
flatbuffers==1.12
future==0.18.2
gast==0.4.0
google-auth==2.0.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.39.0
gym==0.19.0
h5py==3.1.0
idna==3.2
ImageHash==4.2.1
imageio==2.9.0
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
keras==2.6.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
lpips==0.1.3
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mistune==0.8.4
mysql-connector-python==8.0.26
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
notebook==6.4.0
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.3.56
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.0
pandas==1.3.4
pandocfilters==1.4.3
parso==0.8.2
pefile==2021.5.24
pickleshare==0.7.5
Pillow==8.3.1
prometheus-client==0.11.0
prompt-toolkit==3.0.19
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
Pygments==2.9.0
pyinstaller==4.4
pyinstaller-hooks-contrib==2021.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyWavelets==1.1.1
pywin32==301
pywin32-ctypes==0.2.0
pywinpty==1.1.3
pyzmq==22.1.0
qtconsole==5.1.1
QtPy==1.9.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-image==0.18.2
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.2
Send2Trash==1.7.1
six==1.15.0
tb-nightly==2.7.0a20210904
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-cpu==2.6.0
tensorflow-estimator==2.6.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.7.2
torch @ file:///D:/Epan/torch-1.8.0%2Bcu111-cp38-cp38-win_amd64.whl
torchstat==0.0.7
torchvision @ file:///D:/Epan/torchvision-0.9.0%2Bcu111-cp38-cp38-win_amd64.whl
tornado==6.1
tqdm==4.61.2
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
